Move SchemaWithPath to lineage
