+++ src/main/java/no/ssb/dapla/dataset/doc/model/lineage/SchemaWithPath.java
